+++ plan/util.py
@@ -37,18 +37,3 @@ def get_degree_info(plan_title):
 
     return {"degreeType": plan_degree_type, "degree": plan_degree}
 
-
-def get_fields_from_schema(schema):
-    text_fields = []
-    obj_fields = []
-    for property, spec in schema["properties"].items():
-        if spec["type"] == "string":
-            text_fields.append(property)
-        elif spec["type"] in ["array", "object"]:
-            obj_fields.append(property)
-        else:
-            ValueError(
-                "We shouldn't have types other than string, array and object."
-            )
-    all_fields = text_fields + obj_fields
-    return text_fields, obj_fields, all_fields
